private example init
