+++ .gitignore
@@ -24,8 +24,8 @@
 # Generated by Cargo
 Cargo.lock
 
-# Build directory for `cargo build`
-build/
+# # Build directory for `cargo build`
+# build/
 
 # Binaries produced by `cargo install`
 bin/
