+++ Cargo.toml
@@ -1,6 +1,6 @@
 [workspace]
 resolver = "2"
-members = ["cli", "examples/private-input", "hdp"]
+members = ["cli", "examples/private-input-module", "hdp"]
 
 [workspace.package]
 version = "0.4.0"
