+++ examples/private-input-module/private_module/Scarb.lock
@@ -1,0 +1,14 @@
+# Code generated by scarb DO NOT EDIT.
+version = 1
+
+[[package]]
+name = "hdp_cairo"
+version = "0.1.0"
+source = "git+https://github.com/HerodotusDev/hdp-cairo.git#4e00d58246e3f5ea5d1495e914c28c738c6d540f"
+
+[[package]]
+name = "private_module"
+version = "0.1.0"
+dependencies = [
+ "hdp_cairo",
+]
