+++ examples/private-input-module/private_module/Scarb.toml
@@ -1,0 +1,15 @@
+[package]
+name = "private_module"
+version = "0.1.0"
+edition = "2023_11"
+
+# See more keys and their definitions at https://docs.swmansion.com/scarb/docs/reference/manifest.html
+
+[dependencies]
+starknet = "2.6.3"
+hdp_cairo = { git = "https://github.com/HerodotusDev/hdp-cairo.git" }
+
+[[target.starknet-contract]]
+sierra = true
+casm = true
+casm-add-pythonic-hints = true
